fixed disoplaying of newly added stocks, also merged change name for portfolio
