+++ src/components/ui/custom/createPortfolioButton.tsx
@@ -23,6 +23,8 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
+import { AuthContext } from "@/js/AuthContext";
+import { toast } from "sonner";
 
 const formSchema = z.object({
   portfolioName: z
@@ -30,8 +32,11 @@ const formSchema = z.object({
     .min(1, { message: "Name must be at least 1 character long" }),
 });
 
-export function CreatePortfolioButton({ onSubmit, handleCreateButtonClick }) {
+export function CreatePortfolioButton({ onSubmit, portfolios }) {
   const [open, setOpen] = React.useState(false);
+  // AUTH CONTEXT
+  const { subscriptionType } = React.useContext(AuthContext);
+
 
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
@@ -53,9 +58,21 @@ export function CreatePortfolioButton({ onSubmit, handleCreateButtonClick }) {
     setOpen(false);
   };
 
+  // limit amount of portfolios based on subscription type
+  function handleCreateButtonClick() {
+    if (portfolios.length >= 1 && subscriptionType === "FREE") {
+      toast.error("Free users can only have one portfolio.");
+      return false;
+    }
+    else if (portfolios.length >= 10 && subscriptionType === "PREMIUM") {
+      toast.error("Premium users can only have up to 10 portfolios.");
+      return false;
+    }
+  }
+
   return (
     <>
-      
+
       <Dialog open={open} onOpenChange={setOpen}>
         <DialogTrigger asChild>
           <Button onClick={onTriggerClick} >Create Portfolio</Button>
