+++ src/views/portfolioOverview.tsx
@@ -29,6 +29,7 @@ import { createPortfolio, fetchPortfolios, fetchUpdatePortfolioName } from "@/ap
 import { PortfolioEditDialog } from "@/components/ui/custom/portfolioEditDialog";
 import { Pencil } from "lucide-react";
 import { Button } from "@/components/ui/button";
+import SearchBar from "@/components/divex/searchBar";
 
 
 
@@ -36,38 +37,7 @@ export default function PortfolioOverview() {
   // PORTFOLIO STATES
 
   const { portfolios, setPortfolios } = usePortfolios();
-
-  // const [selectedPortfolio, setSelectedPortfolio] = React.useState<Portfolio | null>(() => {
-  //   const stored = localStorage.getItem("selectedPortfolio");
-  //   return stored ? JSON.parse(stored) : null;
-  // });
-  
-  const [selectedPortfolio, setSelectedPortfolio] = useState<Portfolio | null>(null)
-  
-
-  //SEARCH STATES
-  const [searchValue, setSearchValue] = useState("");
-  const [searchResults, setSearchResults] = useState<Stock[]>([])
-
-
-  // AUTH CONTEXT
-  const { subscriptionType } = useContext(AuthContext);
-
-
-  // limit amount of portfolios based on subscription type
-  function handleCreateButtonClick() {
-    if (portfolios.length >= 1 && subscriptionType === "FREE") {
-      toast.error("Free users can only have one portfolio.");
-      return false;
-    }
-    else if (portfolios.length >= 10 && subscriptionType === "PREMIUM") {
-      toast.error("Premium users can only have up to 10 portfolios.");
-      return false;
-    }
-  }
-
-
-
+  const [selectedPortfolio, setSelectedPortfolio] = useState<Portfolio | null>(null);
 
 
   async function handlePortfolioCreation(values) {
@@ -85,106 +55,86 @@ export default function PortfolioOverview() {
       const data = await fetchPortfolios();
       setPortfolios(data || []);
     }
+
     loadPortfolios();
+    setSelectedPortfolio(selectedPortfolio);
   }, []);
 
 
-  useEffect(() => {
-    async function fetchAndSetResults() {
-      try {
-        const results = await fetchStocksBySearchTerm(searchValue);
-        if (results) {
-          setSearchResults(results);
-        }
-      } catch (error) {
-        console.error(error);
-      }
+  const changePortfolioName = async (newName: string) => {
+    const token = localStorage.getItem("token");
+    if (!selectedPortfolio) {
+      toast.error("No portfolio selected.");
+      return;
     }
-
-    if (searchValue) {
-      fetchAndSetResults();
-    } else {
-      setSearchResults([]);
+    if (!token) {
+      toast.error("No token found. Please log in.");
+      return;
     }
-  }, [searchValue]);
+    try {
+      const data = await fetchUpdatePortfolioName(newName, selectedPortfolio.name);
 
+      // Update local state
+      const updatedPortfolio = { ...selectedPortfolio, name: data.name };
+      setSelectedPortfolio(updatedPortfolio);
+      setPortfolios((prevPortfolios) =>
+        prevPortfolios.map((portfolio) =>
+          portfolio.id === updatedPortfolio.id ? updatedPortfolio : portfolio
+        )
+      );
 
-  // Should propably use prop drilling?
-  const navigate = useNavigate();
-  function showStockDetails(ticker: string) {
-    navigate("/stocks/" + ticker, { state: portfolios });
+      // Update localStorage
+      localStorage.setItem("selectedPortfolio", JSON.stringify(updatedPortfolio));
+
+      toast.success("Portfolio name updated.");
+
+    } catch (error: any) {
+      console.error("Update portfolio name error", error);
+      toast.error(error.message);
+    }
   }
 
 
-console.log("portfolios", portfolios)
-console.log("Localstorage Ports", JSON.parse(localStorage.getItem("selectedPortfolio")))
+  console.log("portfolios", portfolios)
+  console.log("Localstorage Ports", JSON.parse(localStorage.getItem("selectedPortfolio")))
 
-console.log("Selected PORTs", selectedPortfolio)
+  console.log("Selected PORTs", selectedPortfolio)
+
+
   return (
     <>
-      <div className="flex flex-row items-center gap-4">
-        <h1 className="text-semibold text-5xl">Portfolio</h1>
-      </div>
-
       <div className="flex flex-row items-center gap-4 mt-5">
 
         {portfolios !== null && (
-          <div className="content-center">
-            <PortfolioSelect
-              portfolioList={portfolios}
-              selectedPortfolio={selectedPortfolio}
-              setSelectedPortfolio={setSelectedPortfolio}
-            />
+          <div className="flex flex-col content-center">
+
+            <div className="relative">
+              <h1 className="text-semibold flex text-5xl">{selectedPortfolio ? selectedPortfolio.name : "Select a portfolio"}
+                <PortfolioEditDialog
+                  onSubmit={changePortfolioName}
+                  selectedPortfolio={selectedPortfolio}
+                />
+              </h1>
+            </div>
+
+            <div className="flex items-center gap-3 mt-4">
+              <PortfolioSelect
+                portfolioList={portfolios}
+                selectedPortfolio={selectedPortfolio}
+                setSelectedPortfolio={setSelectedPortfolio}
+              />
+
+              <div className="ml-1 content-center">
+                <CreatePortfolioButton
+                  onSubmit={handlePortfolioCreation}
+                  portfolios={portfolios}
+                />
+              </div>
+            </div>
           </div>
         )}
 
-        <div className="ml-1 content-center">
-          <CreatePortfolioButton
-            onSubmit={handlePortfolioCreation}
-            handleCreateButtonClick={handleCreateButtonClick}
-          />
-        </div>
-
-        {/* TODO FIX BG COLOR TO MATCH THEMEPROVIDER */}
-        <div className="relative w-full max-w-md ml-auto">
-          <Input
-            id="search"
-            placeholder="Search..."
-            type="search"
-            className="p-4 border-2 border-gray-400 w-full"
-            value={searchValue}
-            onChange={(e) => setSearchValue(e.target.value)}
-          />
-
-          {searchResults.length > 0 && (
-            <div
-              className="absolute left-0 w-full border-2 border-gray-400 bg-white z-50 shadow-lg overflow-y-auto"
-              style={{
-                maxHeight: "calc(100vh - 300px)", // Adjust the dropdown height to fit within the viewport
-              }}
-            >
-              <div className="flex flex-col items-start">
-                {searchResults.map((stock) => (
-                  <div
-                    onClick={() => showStockDetails(stock.ticker)}
-                    key={stock.ticker}
-                    className="w-full p-2 hover:bg-gray-100 cursor-pointer border-b last:border-none"
-                  >
-                    <h1 className="flex justify-start font-semibold">{stock.name}</h1>
-                    <h3 className="flex justify-start text-sm text-gray-500">{stock.ticker}</h3>
-                  </div>
-                ))}
-              </div>
-            </div>
-          )}
-
-          {searchValue && searchResults.length === 0 && (
-            <div className="flex flex-col items-start absolute w-full border-2 border-gray-400 bg-white z-50 shadow-lg p-4">
-              <p>No results found for "{searchValue}".</p>
-            </div>
-          )}
-
-        </div>
+        <SearchBar />
       </div>
 
 
@@ -199,7 +149,6 @@ console.log("Selected PORTs", selectedPortfolio)
 
         {selectedPortfolio !== null && (
           <div className='w-full bg-primary-foreground shadow-md rounded-lg p-6 mt-5'>
-            <h1 className="flex justify-start font-semibold text-2xl ml-1 mb-6">{selectedPortfolio.name}</h1>
 
             {selectedPortfolio?.portfolioEntries?.length === 0 || selectedPortfolio?.portfolioEntries?.length === undefined && (
               <h1 className="text-4xl font-semibold mb-12">Add stocks to the portfolio</h1>
